Refactor the reactor
