+++ lib/process_host.rb
@@ -2,5 +2,6 @@ require 'connection'
 require 'telemetry/logger'
 
 require 'process_host/cooperation'
+require 'process_host/cooperation/reactor/dispatcher'
 require 'process_host/cooperation/reactor'
 require 'process_host/process_host'
